backend/requirements.txt
# FastAPI and ASGI Server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7
alembic==1.12.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Environment Variables
python-dotenv==1.0.0

# Data Validation (included with FastAPI but explicit)
pydantic==2.5.0
pydantic-settings==2.1.0

# CORS (for frontend-backend communication)
python-multipart==0.0.6

# Optional: Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

